+++ CHANGELOG.md
@@ -6,6 +6,8 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 
 ## [Unreleased]
 
+## [5.5.2] - 2021-04-12
+
 ### Added
 
 - [GH-305](https://github.com/llaville/php-compat-info/issues/305) `--exclude` option to be able to filter data source
@@ -153,7 +155,9 @@ using the [Keep a CHANGELOG](http://keepachangelog.com) principles.
 
 - add support to PHP-Parser 3.1 for running on PHP >= 5.5 and for parsing code PHP 5.2 to PHP 7.2
 
-[unreleased]: https://github.com/llaville/php-compat-info/compare/5.5.0...HEAD
+[unreleased]: https://github.com/llaville/php-compat-info/compare/5.5.2...HEAD
+[5.5.2]: https://github.com/llaville/php-compat-info/compare/5.5.1...5.5.2
+[5.5.1]: https://github.com/llaville/php-compat-info/compare/5.5.0...5.5.1
 [5.5.0]: https://github.com/llaville/php-compat-info/compare/5.4.4...5.5.0
 [5.4.4]: https://github.com/llaville/php-compat-info/compare/5.4.3...5.4.4
 [5.4.3]: https://github.com/llaville/php-compat-info/compare/5.4.2...5.4.3